+++ packages/javascript/lib/math/index.ts
@@ -3,5 +3,6 @@ export { div } from './div.js';
 export { mul } from './mul.js';
 export { numPick } from './numPick.js';
 export { numUnique } from './numUnique.js';
+export { round } from './round.js';
 export { sub } from './sub.js';
 export { sum } from './sum.js';